requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# app.py: 2
Pillow == 7.1.2

# app.py: 1
streamlit == 0.67.1
